Moving everything related to changes into own namespace
