+++ Source/Kernel/Events.Projections/EventContext.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Cratis.Events.Projections.Changes;
+
 namespace Cratis.Events.Projections
 {
     /// <summary>
