+++ Source/Kernel/Events.Projections/IProjection.cs
@@ -2,10 +2,10 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Dynamic;
+using Cratis.Events.Projections.Changes;
 
 namespace Cratis.Events.Projections
 {
-
     /// <summary>
     /// Defines a projection.
     /// </summary>
