+++ Source/Kernel/Events.Projections/IProjectionStorage.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Dynamic;
+using Cratis.Events.Projections.Changes;
 
 namespace Cratis.Events.Projections
 {
